Fix duplication bug
Items with metadata that had nested metadata (i.e. of type
ConfigurationSerializable) were not serializing correctly. This meant that
MongoDB attempted to serialize the items itself, which failed. Hence, the
chest did not save and so neither did taken items, causing a duplication
bug.

This issue was that when the serialization function
(ItemStackUtils#serializeMap) got to a ConfigurationSerializable, it did
not *fully* serialize the metadata. Rather, it relied upon Bukkit and
Bukkit alone to serialize it -- a problem which that method in itself
tries to fix.

The fix is to add a recursive function call so that the
'ConfigurationSerializable' is serialized correctly.